[MEET-19] Implemented OAuth2 login, storing user data in session, path for handling JWT. (#11)
* Add Spring Security & google auth

* Added SecurityConfiguration to give unauthorized users access to homepage

* Created a way to retrieve user's attributes from his token (when authorization passed)

* Added Client retrieval via token

* Added HttpSession list holder, tried to test this bs

* Preparing Jwt validation for Postman

* Given a valid JWT HttpSession is being set as authenticated

* Converted Client & Announcement classes to handle user retrieval from session, finished token bearer verifier, user email retrieval from session

* Removed useless HttpSessionConfig, added method to retrieve e-mail from token bearer, fixed mapping in SvelteController

* Fully implemented Service & Controller for the Announcement, added validation using HttpSession info about user

* Made controllers & services for Event & Meeting use HttpSession to verify user data

+++ backend/MeetApp/src/main/java/com/meetapp/meetapp/service/AnnouncementService.java
@@ -1,14 +1,15 @@
 package com.meetapp.meetapp.service;
 
-import com.meetapp.meetapp.model.Announcement;
 import com.meetapp.meetapp.dto.AnnouncementDTO;
+import com.meetapp.meetapp.model.Announcement;
 import com.meetapp.meetapp.model.Client;
 import com.meetapp.meetapp.model.Location;
 import com.meetapp.meetapp.repository.AnnouncementRepository;
+import com.meetapp.meetapp.repository.ClientRepository;
 import com.meetapp.meetapp.repository.LocationRepository;
+import com.meetapp.meetapp.security.SessionManager;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Service;
-
-import lombok.val;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -17,10 +18,13 @@ import java.util.NoSuchElementException;
 public class AnnouncementService {
     private final AnnouncementRepository announcementRepository;
     private final LocationRepository locationRepository;
+    private final ClientRepository clientRepository;
 
-    public AnnouncementService(AnnouncementRepository announcementRepository, LocationRepository locationRepository) {
+    public AnnouncementService(AnnouncementRepository announcementRepository, LocationRepository locationRepository,
+                               ClientRepository clientRepository) {
         this.announcementRepository = announcementRepository;
         this.locationRepository = locationRepository;
+        this.clientRepository = clientRepository;
     }
 
     public List<Announcement> retrieveAnnouncements() {
@@ -31,40 +35,56 @@ public class AnnouncementService {
         return findAnnouncementOrThrow(announcementId);
     }
 
-    public Announcement createAnnouncement(AnnouncementDTO newAnnouncement) {
-        // TODO: Convert token
+    public Announcement createAnnouncement(AnnouncementDTO newAnnouncement, HttpSession sess) {
+        String email = SessionManager.retrieveEmailOrThrow(sess);
+        Client foundClient = findClientOrThrow(email);
+        Location foundLocation = findLocationOrThrow(newAnnouncement.getLocationId());
 
-        val foundLocation = findLocationOrThrow(newAnnouncement.getLocationId());
-
-        // FIXME new Client()
         Announcement announcementToSave =
-                new Announcement(new Client(), foundLocation, newAnnouncement.getDescription(),
+                new Announcement(foundClient, foundLocation, newAnnouncement.getDescription(),
                         newAnnouncement.getTitle());
 
         return announcementRepository.save(announcementToSave);
     }
 
-    public Announcement updateAnnouncement(Integer announcementId, AnnouncementDTO updatedAnnouncement) {
+    public Announcement updateAnnouncement(Integer announcementId, AnnouncementDTO updatedAnnouncement, HttpSession sess) {
+        String email = SessionManager.retrieveEmailOrThrow(sess);
+        Client supposedAuthor = findClientOrThrow(email);
         Location foundLocation = findLocationOrThrow(updatedAnnouncement.getLocationId());
         Announcement foundAnnouncement = findAnnouncementOrThrow(announcementId);
 
-        foundAnnouncement.setLocation(foundLocation);
-        foundAnnouncement.setTitle(updatedAnnouncement.getTitle());
-        foundAnnouncement.setDescription(updatedAnnouncement.getDescription());
-        return announcementRepository.save(foundAnnouncement);
+        if (foundAnnouncement.getAuthor().equals(supposedAuthor)) {
+            foundAnnouncement.setLocation(foundLocation);
+            foundAnnouncement.setTitle(updatedAnnouncement.getTitle());
+            foundAnnouncement.setDescription(updatedAnnouncement.getDescription());
+            return announcementRepository.save(foundAnnouncement);
+        } else {
+            throw new SecurityException("Announcement with id: " + announcementId + " does not belong to the user with id: " + supposedAuthor.getId());
+        }
     }
 
-    public void deleteAnnouncement(Integer announcementId) {
-        announcementRepository.deleteById(announcementId);
+    public void deleteAnnouncement(Integer announcementId, HttpSession sess) {
+        Announcement announcementToDelete = findAnnouncementOrThrow(announcementId);
+        String email = SessionManager.retrieveEmailOrThrow(sess);
+        Client supposedAuthor = findClientOrThrow(email);
+
+        if (announcementToDelete.getAuthor().equals(supposedAuthor)) {
+            announcementRepository.deleteById(announcementId);
+        }
     }
 
-    Announcement findAnnouncementOrThrow(Integer announcementId) {
+    public Announcement findAnnouncementOrThrow(Integer announcementId) {
         return announcementRepository.findById(announcementId).orElseThrow(
                 () -> new NoSuchElementException("An announcement with id: " + announcementId + " does not exist."));
     }
 
-    Location findLocationOrThrow(Integer locationId) {
+    public Location findLocationOrThrow(Integer locationId) {
         return locationRepository.findById(locationId).orElseThrow(
                 () -> new NoSuchElementException("A location with id: " + locationId + " does not exist."));
+    }
+
+    public Client findClientOrThrow(String email) {
+        return clientRepository.findClientByEmail(email).orElseThrow(
+                () -> new NoSuchElementException("A client with email: " + email + " does not exist."));
     }
 }
